requirements.txt
annotated-types==0.7.0
anyio==4.10.0
boto3==1.40.6
botocore==1.40.6
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
duckdb==1.3.2
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
fsspec==2025.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
mysql-connector-python==9.4.0
numpy==2.3.2
pandas==2.3.1
polars==1.32.0
#pyarrow==21.0.0
#pydantic==2.11.7
#pydantic_core==2.33.2
#Pygments==2.19.2
##pyiceberg==0.9.1
#pyparsing==3.2.3
#python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==13.9.4
rich-toolkit==0.14.9
rignore==0.6.4
s3transfer==0.13.1
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.47.2
strictyaml==1.7.3
tabulate==0.9.0
tenacity==9.1.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
